Avoid unnecessary BigDecimal/float to double conversions

+++ order-service/src/main/java/ionutbalosin/training/ecommerce/order/dao/mapper/OrderRowMapper.java
@@ -25,7 +25,7 @@ public class OrderRowMapper implements RowMapper<Order> {
         .id(UUID.fromString(rs.getString(Order.ID)))
         .sourceEventId(UUID.fromString(rs.getString(Order.SOURCE_EVENT_ID)))
         .userId(UUID.fromString(rs.getString(Order.USER_ID)))
-        .amount(rs.getFloat(Order.AMOUNT))
+        .amount(rs.getDouble(Order.AMOUNT))
         .currency(rs.getString(Order.CURRENCY))
         .details(stringToJsonObject(rs.getString(Order.DETAILS)))
         .status(fromValue(rs.getString(Order.STATUS)))
