+++ shopping-cart-service/src/main/java/ionutbalosin/training/ecommerce/shopping/cart/model/mapper/CartItemMapper.java
@@ -19,7 +19,7 @@ public class CartItemMapper {
         .userId(userId)
         .productId(cartItemCreate.getProductId())
         .quantity(cartItemCreate.getQuantity())
-        .discount(cartItemCreate.getDiscount().floatValue())
+        .discount(cartItemCreate.getDiscount())
         .dateIns(LocalDateTime.now())
         .usrIns("anonymous")
         .stat("A");
